Add input boxes to write custom segments.
This commit also fixes some issues with the spinner direction and
how segments are drawn. This is so that we can detect the spinner
location amongst the segments. The change was to make sure the top of
the spinner registers as 0.0 and the bottom as Math.PI radians. This
makes it easier to pick the correct segment by calculating a proportion
of the circumference of the spinner that the pointer has gone past.